# src/main/java/br/com/pauloneto/medicalsystem/Usuario/Domain/Usuario.java
package br.com.pauloneto.medicalsystem.Usuario.Domain;

import jakarta.persistence.*;
import lombok.Data;

import java.util.Date;

@Data
@Entity
@Table(name = "USUARIOS")
public class Usuario {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "ID_USUARIO")
    private Long idUsuario;
    @Column(name = "NOME_USUARIO")
    private String nomeUsuario;
    @Column(name = "EMAIL")
    private String email;
    @Column(name = "CPF")
    private String cpf;
    @Column(name = "TELEFONE")
    private String telefone;
    @Column(name = "DATA_NASCIMENTO")
    private Date dataNascimento;
    @Column(name = "PERMISSAO")
    private Permissao permissao;

    public Usuario(Long idUsuario, String nomeUsuario, String email, String cpf, String telefone, Date dataNascimento,Permissao permissao){
        this.idUsuario = idUsuario;
        this.nomeUsuario = nomeUsuario;
        this.email = email;
        this.cpf = cpf;
        this.telefone = telefone;
        this.dataNascimento = dataNascimento;
        this.permissao = permissao;



    }

    public Usuario() {

    }

    public enum Permissao{
        ADMIN("ADMIN"),
        SECRETARIO("SECRETARIO");

        private final String descricao;

        Permissao(String descricao) {
            this.descricao = descricao;

        }

        public String getDescricao(){
            return descricao;
        }



    }
}